% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateBzones.R
\docType{data}
\name{AveHhPerBlockGroup}
\alias{AveHhPerBlockGroup}
\title{Average households per block group}
\format{A number:
\describe{
 \item{AveHhPerBlockGroup}{average households per census block group}
}}
\source{
CreateBzones.R script.
}
\usage{
AveHhPerBlockGroup
}
\description{
A number representing the average number of households per census block.
}
\keyword{datasets}

