% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateBzoneDev.R, R/CreateBzones.R, R/CreateHouseholds.R
\name{buildModule}
\alias{buildModule}
\title{buildModule' function that, when run, creates an environment that contains all
CreateBzoneDevModule'. Then a function named 'CreateBzoneDev' is defined that}
\usage{
buildModule()

buildModule()

buildModule()
}
\description{
buildModule' function that, when run, creates an environment that contains all
CreateBzoneDevModule'. Then a function named 'CreateBzoneDev' is defined that

\code{buildModule} builds the module that will carry out the demo models.

\code{buildModule} builds the module that will carry out the demo models.
}
\details{
This function builds the module when it is run. The module is created as an
environment which contains all of the necessary elements.

This function builds the module when it is run. The module is created as an
environment which contains all of the necessary elements.
}

