% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateBzoneDev.R
\name{calcBuildingTypes}
\alias{calcBuildingTypes}
\title{Simulate Bzone dwelling types.}
\usage{
calcBuildingTypes(L)
}
\arguments{
\item{L}{A list containing the following components that have been either
read from the datastore or produced by the application of the
calcBzoneDistDen function:
PopDen: A numeric vector of the population density of each Bzone produced by
the calcBzoneDistDen function.
NumHh: An integer vector of the number of households in each Bzone read from
the Bzone table.}
}
\value{
A list containing the following components:
SfdNum: An integer vector of the number of households living in single-family
dwellings in each Bzone.
MfdNum: An integer vector of the number of households living in multifamily
dwellings in each Bzone.
}
\description{
\code{calcBuildingTypes} calculates the number of single-family dwellings and
the number of multifamily dwelling units in each Bzone.
}
\details{
This function calculates the number of single-family detached dwellings and
multifamily dwellings in each Bzone. This is done using a toy model which
relates the proportion of multifamily dwellings to population density. The
results are the numbers of single-family detached dwellings and multifamily
detached dwellings in each Bzone.
}

