% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateBzoneDev.R
\name{calcBzoneDistDen}
\alias{calcBzoneDistDen}
\title{Simulate Bzone density and distance characteristics}
\usage{
calcBzoneDistDen(L)
}
\arguments{
\item{L}{A list containing the following components that have been read
from the datastore:
Bzone: A character vector of Bzone names read from the Bzone table.
DevType: A character vector identifying the development type of each Bzone
read from the Bzone table.
NumHh: A integer vector of the number of households in each Bzone read from
the Bzone table.
Area: A number identifying the geographic area of the metropolitan area read
from the Marea table.}
}
\value{
A list containing the following components:
DistFromCtr: A numeric vector of the distance of each Bzone from the urban
area center.
PopDen: A numeric vector of the population density of each Bzone.
}
\description{
\code{calcBzoneDistDen} simulates the population density of all Bzones
and distance from the urban core for Bzones whose DevType is "Metropolitan".
}
\details{
The Bzone population density and distance from the urban core for
metropolitan Bzones are simulated using population, population density, and
distance distributions compiled by the U.S. Census Bureau.
}

