% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateHouseholds.R
\name{calcHhSizeProp}
\alias{calcHhSizeProp}
\title{Calculate proportions of households by household size}
\usage{
calcHhSizeProp(Inp_ls = HouseholdSizesInp_ls)
}
\arguments{
\item{Inp_ls}{A list containing the specifications for
the estimation data contained in "household_sizes.csv".}
}
\value{
A data frame having two columns names "Size" and "Proportion" that
contains data on the proportions of households in households having sizes
between 1 and 7 persons per household.
}
\description{
\code{calcHhSizeProp} calculates the proportions of households by household
size from the number of households by household size.
}
\details{
This function produces a data frame listing the proportions of households
by household size from a data frame that lists the numbers of households by
household size. The rows of the data frame are put in order of ascending
household size.
}

