% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateBzoneDev.R
\name{CreateBzoneDev}
\alias{CreateBzoneDev}
\title{Build the module.}
\usage{
CreateBzoneDev()
}
\description{
\code{CreateBzoneDev} produces the module that will simulate various
development characteristics of Bzones.
}
\details{
This function produces the module that will simulate various Bzone
development characteristics when it is run.
}

