% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateBzones.R
\name{CreateBzones}
\alias{CreateBzones}
\title{Creates a set of simulated Bzones}
\usage{
CreateBzones(L)
}
\arguments{
\item{L}{A list containing the following components that have been read
from the datastore:
Azone: A character vector of Azone names read from the Azone table.
NumHh: A numeric vector of the number of people in each Azone read from the
Azone table.
Marea: A character vector of Marea names read from the Azone table.
Metropolitan: A numeric vector of the proportion of households that are of
the Metropolitan development type in each Azone read from the Azone table.
Town: A numeric vector of the proportion of households that are of the Town
development type in each Azone read from the Azone table.
Rural: A numeric vector of the proportion of households that are of the
Rural development type in each Azone read from the Azone table.}
}
\value{
A list containing the following components:
Bzone: A character vector of the names of the Bzones that is to be assigned
to the Bzone table.
Azone: A character vector of the names of the Azones associated with each
Bzone that is to be assigned to the Bzone table.
Marea: A character vector of the names of the Mareas associated with each
Bzone that is to be assigned to the Bzone table.
DevType: A character vector identifying the development type of each Bzone
that is to be assigned to the Bzone table.
NumHh: An integer vector identifying the number of households assigned to
each Bzone that is to be assigned to the Bzone table.
LENGTH: A named integer vector having a single named element, "Bzone",
which identifies the length (number of rows) of the Bzone table to be
created in the datastore.
SIZE: a named numeric vector having four elements. The first element,
"Azone", identifies the size of the longest Azone name. The second element,
"Bzone", identifies the size of the longest Bzone name. The third element,
"Marea", identifies the size of the longest Marea name. The fourth element,
"DevType", identifies the size of the longest DevType name.
}
\description{
\code{CreateBzones} creates a set of simulated Bzones and assigns
development types, Azones, Mareas, unique Bzone IDs, and numbers of
households.
}
\details{
This function creates a set of simulated Bzones and assigns attributes
including the Azone that the Bzone is situated in, the Marea that the Bzone
is situated in, the number of households in the Bzone, and the development
type (i.e. Metropolitan, Town, Rural) of the Bzone. These data items are
vectors that are to be stored in the "Bzone" table. Since this table does not
exist, the function calculates a LENGTH value for the table and returns that
as well. The framework uses this information to initialize the Bzone table.
The function also computes the maximum numbers of characters in the Bzone,
Azone, Marea, and DevType datasets and assigns these to a SIZE vector. This
is necessary so that the framework can initialize these datasets in the
datastore. All results are returned in a list.
}

