% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateHouseholds.R
\name{CreateHouseholds}
\alias{CreateHouseholds}
\title{Creates a set of simulated households}
\usage{
CreateHouseholds(L)
}
\arguments{
\item{L}{A list containing the following components:
Azone: A character vector of Azone names read from the Azone table.
Population: A numeric vector of the number of people in each Azone read from
the Azone table.}
}
\value{
A list containing the following components:
HhSize: An integer vector of the calculated sizes of the simulated households
that is to be assigned to the Household table.
Azone: A character vector of the names of the Azones the simulated households
are assigned to that is to be assigned to the Household table.
HhId: A character vector identifying the unique ID for each household that is
to be assigned to the Household table.
NumHh: An integer vector identifying the number of households assigned to
each Azone that is to be assigned to the Azone table.
LENGTH: A named integer vector having a single named element, "Household",
which identifies the length (number of rows) of the Household table to be
created in the datastore.
SIZE: A named integer vector having two elements. The first element, "Azone",
identifies the size of the longest Azone name. The second element, "HhId",
identifies the size of the longest HhId.
}
\description{
\code{CreateHouseholds} creates a set of simulated households that each have
a unique household ID, an Azone to which it is assigned, and a household
size (number of people in the household).
}
\details{
This function creates a set of simulated households where each household is
assigned a household size, an Azone, and a unique ID. These data items are
vectors that are to be stored in the "Household" table. Since this table does
not exist, the function calculates a LENGTH value for the table and returns
that as well. The framework uses this information to initialize the
Households table. The function also computes the maximum numbers of
characters in the HhId and Azone datasets and assigns these to a SIZE vector.
This is necessary so that the framework can initialize these datasets in the
datastore. All the results are returned in a list.
}

