% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateHouseholds.R
\docType{data}
\name{HhSizeProp_df}
\alias{HhSizeProp_df}
\title{Household size proportions}
\format{A data frame with 7 rows and 2 variables:
\describe{
 \item{Size}{household size}
 \item{Proportion}{proportion of households in household size category}
}}
\source{
CreateHouseholds.R script.
}
\usage{
HhSizeProp_df
}
\description{
A dataset containing the proportions of households by household size.
}
\keyword{datasets}

