% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateBzoneDev.R
\docType{data}
\name{MiscDevParameters_df}
\alias{MiscDevParameters_df}
\title{Miscellaneous Bzone development model parameters}
\format{A data frame with 3 rows and 2 variables:
\describe{
 \item{Parameter}{name of the parameter}
 \item{Value}{value of the parameter}
}}
\source{
CreateBzonesDev.R script.
}
\usage{
MiscDevParameters_df
}
\description{
A dataset containing parameters for average household size, average town
population density, and average rural population density.
}
\keyword{datasets}

