% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateBzoneDev.R
\docType{data}
\name{PopByDistance_df}
\alias{PopByDistance_df}
\title{Population proportions by distance}
\format{A data frame with 40 rows and 2 variables:
\describe{
 \item{Distance}{distance from urban center, in miles}
 \item{PopProp}{proportion of urbanized area population}
}}
\source{
CreateBzonesDev.R script.
}
\usage{
PopByDistance_df
}
\description{
A dataset containing the proportions of urbanized area population by distance
from the urban center rounded to nearest whole mile.
}
\keyword{datasets}

