% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateBzoneDev.R
\docType{data}
\name{PopDenByDistance_df}
\alias{PopDenByDistance_df}
\title{Population density by distance}
\format{A data frame with 40 rows and 2 variables:
\describe{
 \item{Distance}{distance from urban center, in miles}
 \item{PopDensity}{population density, in persons per square mile}
}}
\source{
CreateBzonesDev.R script.
}
\usage{
PopDenByDistance_df
}
\description{
A dataset containing the average census block group population density by
distance from the urban center rounded to nearest whole mile.
}
\keyword{datasets}

