% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateBzoneDev.R
\name{CreateBzoneDev}
\alias{CreateBzoneDev}
\title{Calculate Bzone development characteristics}
\usage{
CreateBzoneDev(L)
}
\arguments{
\item{L}{A list containing the following components that have been read from
the datastore:
Bzone: A character vector of Bzone names read from the Bzone table.
DevType: A character vector identifying the development type of each Bzone
read from the Bzone table.
NumHh: A integer vector of the number of households in each Bzone read from
the Bzone table.
Area: A number identifying the geographic area of the metropolitan area read
from the Marea table.}
}
\value{
A list containing the following components:
DistFromCtr: A numeric vector of the distance of each Bzone from the urban
area center.
PopDen: A numeric vector of the population density of each Bzone.
SfdNum: An integer vector of the number of households living in single-family
dwellings in each Bzone.
MfdNum: An integer vector of the number of households living in multifamily
dwellings in each Bzone.
}
\description{
\code{CreateBzoneDev} is the main function of the CreateBzoneDev module that
calls the calcBzoneDistDen and calcBuildingTypes functions to calculate
Bzone development characteristics.
}
\details{
This is the main function for the CreateBzoneDev module. It is a wrapper for
the calcBzoneDistDen and calcBuildingTypes functions which do all of the
work of calculating Bzone development characteristics including distance from
the urban area center, population density, number of households in single-
family dwellings, and number of households in multifamily dwellings.
}

