% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateBzones.R
\name{CreateBzones}
\alias{CreateBzones}
\title{Build the module.}
\usage{
CreateBzones()
}
\description{
\code{CreateBzones} builds the module that will carry out the creation of Bzones.
}
\details{
This function builds the module when it is run. The module is created as an
environment which contains all of the necessary elements.
}

