% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateHouseholds.R
\name{CreateHouseholds}
\alias{CreateHouseholds}
\title{Build the module.}
\usage{
CreateHouseholds()
}
\description{
\code{CreateHouseholds} builds the module that will carry out the demo models.
}
\details{
This function builds the module when it is run. The module is created as an
environment which contains all of the necessary elements.
}

